Persist job strategies
